Refactor units into its own file
