+++ src/hammer-vlsi/hammer_vlsi_impl.py
@@ -28,6 +28,8 @@ from hammer_logging import HammerVLSILoggingContext
 
 from utils import add_lists, in_place_unique, reverse_dict
 
+from units import TimeValue
+
 
 HammerStepFunction = Callable[['HammerTool'], bool]
 
@@ -150,6 +152,7 @@ HammerToolHookAction = NamedTuple('HammerToolHookAction', [
 
 import hammer_tech
 
+
 class HammerVLSISettings:
     """
     Static class which holds global hammer-vlsi settings.
@@ -175,86 +178,6 @@ class HammerVLSISettings:
         else:
             cls.hammer_vlsi_path = os.environ["HAMMER_VLSI"]
             return True
-
-
-class TimeValue:
-    """Time value - e.g. "4 ns".
-    Parses time values from strings.
-    """
-
-    # From https://stackoverflow.com/a/10970888
-    _prefix_table = {
-        'y': 1e-24,  # yocto
-        'z': 1e-21,  # zepto
-        'a': 1e-18,  # atto
-        'f': 1e-15,  # femto
-        'p': 1e-12,  # pico
-        'n': 1e-9,   # nano
-        'u': 1e-6,   # micro
-        'm': 1e-3,   # mili
-        'c': 1e-2,   # centi
-        'd': 1e-1,   # deci
-        'k': 1e3,    # kilo
-        'M': 1e6,    # mega
-        'G': 1e9,    # giga
-        'T': 1e12,   # tera
-        'P': 1e15,   # peta
-        'E': 1e18,   # exa
-        'Z': 1e21,   # zetta
-        'Y': 1e24,   # yotta
-    }
-
-    def __init__(self, value: str, default_prefix: str = 'n') -> None:
-        """Create a time value from parsing the given string.
-        Default prefix: ns
-        """
-        import re
-
-        regex = r"^([\d.]+) *(.*)s$"
-        m = re.search(regex, value)
-        if m is None:
-            try:
-                num = str(float(value))
-                prefix = default_prefix
-            except ValueError:
-                raise ValueError("Malformed time value %s" % (value))
-        else:
-            num = m.group(1)
-            prefix = m.group(2)
-
-        if num.count('.') > 1 or len(prefix) > 1:
-            raise ValueError("Malformed time value %s" % (value))
-
-        if prefix not in self._prefix_table:
-            raise ValueError("Bad prefix for %s" % (value))
-
-        self._value = float(num) # type: float
-        # Preserve the prefix too to preserve precision
-        self._prefix = self._prefix_table[prefix] # type: float
-
-    @property
-    def value(self) -> float:
-        """Get the value of this time value."""
-        return self._value * self._prefix
-
-    def value_in_units(self, prefix: str, round_zeroes: bool = True) -> float:
-        """Get this time value in the given prefix. e.g. "ns"
-        """
-        retval = self._value * (self._prefix / self._prefix_table[prefix[0]])
-        if round_zeroes:
-            return round(retval, 2)
-        else:
-            return retval
-
-    def str_value_in_units(self, prefix: str, round_zeroes: bool = True) -> str:
-        """Get this time value in the given prefix but including the units.
-        e.g. return "5 ns".
-
-        :param prefix: Prefix for the resulting value - e.g. "ns".
-        :param round_zeroes: True to round 1.00000001 etc to 1 within 2 decimal places.
-        """
-        # %g removes trailing zeroes
-        return "%g" % (self.value_in_units(prefix, round_zeroes)) + " " + prefix
 
 
 ClockPort = NamedTuple('ClockPort', [
@@ -365,6 +288,7 @@ class PlacementConstraint(NamedTuple('PlacementConstraint', [
             }})
         return output
 
+
 class MMMCCornerType(Enum):
     Setup = 1
     Hold = 2
@@ -380,6 +304,7 @@ class MMMCCornerType(Enum):
             return MMMCCornerType.Extra
         else:
             raise ValueError("Invalid mmmc corner type '{}'".format(s))
+
 
 MMMCCorner = NamedTuple('MMMCCorner', [
     ('name', str),
